fix: remove unused normalize_component_type import
