+++ tests/functional/fixtures/inventory.csv
