+++ tests/functional/fixtures/minimal_project/minimal.kicad_pcb
