+++ tests/functional/fixtures/minimal_project/minimal.kicad_sch
