+++ tests/functional/fixtures/modern_project/project.kicad_pcb
