+++ tests/functional/test_functional_base.py
@@ -38,7 +38,7 @@ class FunctionalTestBase(unittest.TestCase):
     def setUp(self):
         """Create temporary directory for test output."""
         # Ensure src is on path
-        src_path = Path(__file__).parent.parent / "src"
+        src_path = Path(__file__).parent.parent.parent / "src"
         if str(src_path) not in sys.path:
             sys.path.insert(0, str(src_path))
