+++ tests/functional/test_integration_all_interfaces.py
