+++ tests/unit/test_api_v3.py
@@ -6,7 +6,7 @@ Tests the new generate_bom() and generate_pos() functions with input=/output= pa
 import sys
 from pathlib import Path
 
-sys.path.insert(0, str(Path(__file__).parent.parent / "src"))
+sys.path.insert(0, str(Path(__file__).parent.parent.parent / "src"))
 
 import unittest
 import tempfile
