+++ tests/unit/test_federated_inventory.py
