+++ tests/unit/test_generic_fabricator.py
