+++ tests/unit/test_project_inventory.py
